last digit

+++ functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  *
  * Return: 0
  */
-int print_last_digit(int n)
+int print_last_digit(long n)
 {
 	if (n < 0)
 		n *= -1;
